Don't allow editing name and deleting the defaults

+++ app/forms/mutations/deleteForm.ts
@@ -7,8 +7,15 @@ const DeleteForm = z.object({
 })
 
 export default resolver.pipe(resolver.zod(DeleteForm), resolver.authorize(), async ({ id }) => {
-  await db.formQuestion.deleteMany({ where: { formId: id } })
-  const form = await db.form.delete({ where: { id } })
+  const form = await db.form.findUnique({ where: { id } })
 
-  return form
+  // Don't allow deleting the Default Form
+  if (form?.name?.toLowerCase() === "default") {
+    throw new Error("Cannot delete the 'Default' form")
+  }
+
+  await db.formQuestion.deleteMany({ where: { formId: id } })
+  const deletedForm = await db.form.delete({ where: { id } })
+
+  return deletedForm
 })
